+++ src/niquests/async_session.py
@@ -71,7 +71,7 @@ from .models import (
 from .packages.urllib3 import ConnectionInfo
 from .packages.urllib3.contrib.resolver._async import AsyncBaseResolver
 from .packages.urllib3.contrib.webextensions._async import load_extension
-from .sessions import Session
+from .sessions import Session, merge_hooks
 from .structures import AsyncQuicSharedCache, CaseInsensitiveDict
 from .utils import (
     _deepcopy_ci,
@@ -133,6 +133,7 @@ class AsyncSession(Session):
         base_url: str | None = None,
         timeout: TimeoutType | None = None,
         headers: HeadersType | None = None,
+        hooks: AsyncHookType[PreparedRequest | Response | AsyncResponse] | None = None,
         revocation_configuration: RevocationConfiguration | None = DEFAULT_STRATEGY,
         app: ASGIApp | None = None,
     ):
@@ -165,7 +166,11 @@ class AsyncSession(Session):
         self.proxies: ProxyType = {}
 
         #: Event-handling hooks.
-        self.hooks: AsyncHookType[PreparedRequest | Response | AsyncResponse] = default_hooks()  # type: ignore[assignment]
+        self.hooks: AsyncHookType[PreparedRequest | Response | AsyncResponse] = (
+            merge_hooks(default_hooks(), hooks)  # type: ignore[assignment]
+            if hooks is not None
+            else default_hooks()
+        )
 
         #: Dictionary of querystring data to attach to each
         #: :class:`Request <Request>`. The dictionary values may be lists for